+++ Tests/DependencyInjection/ConfigurationTest.php
@@ -61,12 +61,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
-                    'view' => [
-                        'default_engine' => null,
-                    ],
                 ],
             ]
         );
@@ -94,12 +88,6 @@ class ConfigurationTest extends TestCase
             ],
             'routing_loader' => false,
             'body_listener' => false,
-            'service' => [
-                'templating' => null,
-            ],
-            'view' => [
-                'default_engine' => null,
-            ],
         ];
 
         $config = $this->processor->processConfiguration($this->configuration, [$config]);
@@ -133,12 +121,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
-                    'view' => [
-                        'default_engine' => null,
-                    ],
                 ],
             ]
         );
@@ -161,12 +143,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
-                    'view' => [
-                        'default_engine' => null,
-                    ],
                 ],
             ]
         );
@@ -190,12 +166,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
-                    'view' => [
-                        'default_engine' => null,
-                    ],
                 ],
             ]
         );
@@ -225,12 +195,6 @@ class ConfigurationTest extends TestCase
                     ],
                     'routing_loader' => false,
                     'body_listener' => false,
-                    'service' => [
-                        'templating' => null,
-                    ],
-                    'view' => [
-                        'default_engine' => null,
-                    ],
                 ],
                 [
                     'format_listener' => [